trunk/arduino/ (saulius@koala.ibt.lt) We need long ints to count correctly :).
git-svn-id: svn+ssh://saulius-grazulis.lt/home/saulius/svn-repositories/microcontrollers/trunk@250 931b4fa1-5130-4491-9317-3c5e5f8f39d3

+++ arduino/Arduino_Uno_74HC4094_FS2012_flow_meter/Arduino_Uno_74HC4094_FS2012_flow_meter.ino
@@ -54,13 +54,13 @@ void setup()
   pinMode(OE,OUTPUT);
 }
 
-unsigned int counter = 0;
+unsigned long long counter = 0;
 
 void loop()
 {
   digitalWrite(OE,HIGH);
   /* Update the display with the current counter value */
-  unsigned int tens_power = 100;
+  unsigned long long tens_power = 100;
   for( int i = 0; i < 8; i++ ) {
     WriteNumberToSegment( i, (counter/tens_power)%10 );
     tens_power *= 10;
